Split Share into provider/controller classes
